refactor(svm): restructure utils into modular files under src folder for better organization
